requirements.txt
appdirs==1.4.3
certifi==2018.1.18
Cython==0.23
Django==1.10.6
docutils==0.13.1
packaging==16.8
pipenv==11.8.0
psutil==5.0.1
pyowm==2.8.0
Pygments==2.2.0
pyparsing==2.2.0
pyTelegramBotAPI==3.6.6
python-telegram-bot==7.0.1
requests==2.13.0
six==1.10.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
beautifulsoup4==4.8.2
Flask==1.1.1
matplotlib==3.1.3
psycopg2-binary==2.8.4
urllib3==1.25.7
COVID19Py==0.3.0
SQLAlchemy==1.3.15
lxml==4.5.0
covid==2.2.6
gunicorn
